requirements.txt
attrs==22.1.0
aws-cdk-lib==2.53.0
aws-cdk.asset-awscli-v1==2.2.26
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.32
aws-cdk.aws-kinesisfirehose-alpha==2.53.0a0
aws-cdk.aws-kinesisfirehose-destinations-alpha==2.53.0a0
cattrs==22.2.0
constructs==10.1.178
exceptiongroup==1.0.4
jsii==1.71.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.4.0
